oct222025

+++ valmerabanicoruperez/src/main/java/com/appdev/vabara/valmerabanicoruperez/repository/RatingRepository.java
@@ -3,5 +3,5 @@ package com.appdev.vabara.valmerabanicoruperez.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.appdev.vabara.valmerabanicoruperez.entity.RatingEntity;
 
-public interface RatingRepository extends JpaRepository<RatingEntity, Integer> {
+public interface RatingRepository extends JpaRepository<RatingEntity, Long> {
 }
